+++ pkg/ffuf/optionsparser.go
@@ -73,6 +73,8 @@ type GeneralOptions struct {
 	ShowVersion               bool     `toml:"-" json:"-"`
 	StopOn403                 bool     `json:"stop_on_403"`
 	StopOnAll                 bool     `json:"stop_on_all"`
+	PauseCode                 string   `json:"pausecode"`
+	PauseTime                 string   `json:"pausetime"`
 	StopOnErrors              bool     `json:"stop_on_errors"`
 	Threads                   int      `json:"threads"`
 	Verbose                   bool     `json:"verbose"`
@@ -93,6 +95,7 @@ type InputOptions struct {
 }
 
 type OutputOptions struct {
+	AuditLog            string `json:"audit_log"`
 	DebugLog            string `json:"debug_log"`
 	OutputDirectory     string `json:"output_directory"`
 	OutputFile          string `json:"output_file"`
@@ -148,6 +151,8 @@ func NewConfigOptions() *ConfigOptions {
 	c.General.StopOn403 = false
 	c.General.StopOnAll = false
 	c.General.StopOnErrors = false
+	c.General.PauseCode = ""
+	c.General.PauseTime = ""
 	c.General.Threads = 40
 	c.General.Verbose = false
 	c.HTTP.Data = ""
@@ -184,6 +189,7 @@ func NewConfigOptions() *ConfigOptions {
 	c.Matcher.Status = "200-299,301,302,307,401,403,405,500"
 	c.Matcher.Time = ""
 	c.Matcher.Words = ""
+	c.Output.AuditLog = ""
 	c.Output.DebugLog = ""
 	c.Output.OutputDirectory = ""
 	c.Output.OutputFile = ""
@@ -562,6 +568,7 @@ func ConfigFromOptions(parseOpts *ConfigOptions, ctx context.Context, cancel con
 	conf.InputNum = parseOpts.Input.InputNum
 
 	conf.InputShell = parseOpts.Input.InputShell
+	conf.AuditLog = parseOpts.Output.AuditLog
 	conf.OutputFile = parseOpts.Output.OutputFile
 	conf.OutputDirectory = parseOpts.Output.OutputDirectory
 	conf.OutputSkipEmptyFile = parseOpts.Output.OutputSkipEmptyFile
@@ -572,6 +579,8 @@ func ConfigFromOptions(parseOpts *ConfigOptions, ctx context.Context, cancel con
 	conf.StopOn403 = parseOpts.General.StopOn403
 	conf.StopOnAll = parseOpts.General.StopOnAll
 	conf.StopOnErrors = parseOpts.General.StopOnErrors
+	conf.PauseCode = parseOpts.General.PauseCode
+	conf.PauseTime = parseOpts.General.PauseTime
 	conf.FollowRedirects = parseOpts.HTTP.FollowRedirects
 	conf.Raw = parseOpts.HTTP.Raw
 	conf.Recursion = parseOpts.HTTP.Recursion